+++ contact/contact.js
@@ -1,0 +1,5 @@
+function showhide(id) {
+    'use strict';
+    var e = document.getElementById(id);
+	e.style.display = (e.style.display == "block") ? "none" : "block";
+}
