--- contact/style.css
@@ -1,3 +1,0 @@
-#contact_form_addon {
-    padding:15px;
-}
